Make prettier-rpc-XXX.js require()able as a module (exposing prettier)
This allows prettier to continue to be runnable from CLI as an RPC server while making it so that when require()d it exposes prettier, the JS module. Rollup defeats most of the common approaches for doing this (as it converts your use of `global`, `module, and `require` to mocks that don't contain enough data to actually detect how the JS file was run. I had to add a hack to get at the actual global object in Node from the rollup'd JS file. Even with that hack, the global object contained way less information when run from eslint for some reason, so I had to special case that.

I don't necessarily think this is a good idea. An alternative would be to just add another rollup config that produces a second bundle that exposes prettier as a JS module matching the JS module used by the RPC server. That might be a lot cleaner.